requirements.txt
PyMuPDF>=1.23.0
pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pytesseract>=0.3.0